pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rlm-mhc"
version = "0.1.0"
description = "Recursive Language Model with Manifold-Constrained Hyper-Connections"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "RLM-mHC Team" }
]
keywords = ["nlp", "transformer", "deep-learning", "recursive-language-model"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "tokenizers>=0.14.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.15.0",
    "einops>=0.7.0",
    "safetensors>=0.4.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "ipython>=8.0.0",
]
gpu = [
    "flash-attn>=2.0.0",
]

[project.urls]
Repository = "https://github.com/rlm-mhc/rlm-mhc"

[project.scripts]
rlm-train = "rlm_mhc.cli:train"
rlm-eval = "rlm_mhc.cli:evaluate"
rlm-generate = "rlm_mhc.cli:generate"

[tool.hatch.build.targets.wheel]
packages = ["src/rlm_mhc"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
